Relocated audit package to root level
