+++ service_packs/kubernetes/iam/helpers.go
@@ -7,7 +7,7 @@ import (
 	"log"
 
 	aibv1 "github.com/Azure/aad-pod-identity/pkg/apis/aadpodidentity"
-	"github.com/citihub/probr/internal/audit"
+	"github.com/citihub/probr/audit"
 	"github.com/citihub/probr/internal/config"
 	"github.com/citihub/probr/service_packs/coreengine"
 	"github.com/citihub/probr/service_packs/kubernetes"
