+++ service_packs/kubernetes/kube.go
@@ -12,7 +12,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/citihub/probr/internal/audit"
+	"github.com/citihub/probr/audit"
 	"github.com/citihub/probr/internal/config"
 	"github.com/citihub/probr/internal/utils"
