+++ service_packs/kubernetes/kubernetes.go
@@ -8,7 +8,7 @@ import (
 	"github.com/markbates/pkger"
 	apiv1 "k8s.io/api/core/v1"
 
-	"github.com/citihub/probr/internal/audit"
+	"github.com/citihub/probr/audit"
 	"github.com/citihub/probr/internal/config"
 	"github.com/citihub/probr/internal/utils"
 	"github.com/citihub/probr/service_packs/coreengine"
