+++ service_packs/kubernetes/kubernetes_mock.go
@@ -1,7 +1,7 @@
 package kubernetes
 
 import (
-	"github.com/citihub/probr/internal/audit"
+	"github.com/citihub/probr/audit"
 	"github.com/stretchr/testify/mock"
 	apiv1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
